Dynamic endpoint auth implementation.
Sample dynamic auth for a the projects findAll endpoint using a
component's method (executing it through the @PreAuthorized Spring
annotation. It shows how we can run Java code from an annotation and
also shows a sample logic to validate the current user roles another
list of roles (it should be feeded with a db's table).